Remove `include/grpcpp/impl/codegen/proto_buffer_reader.h`
